+++ packages/dashboard/src/components/Section/Section.tsx
@@ -16,13 +16,14 @@ export class Section extends React.PureComponent<ISectionProps, ISectionState> {
 
   public render(): JSX.Element {
     const getClassNames = classNamesFunction<ISectionStyleProps, ISectionStyles>();
-    const { disabled, rowHeight } = this.props;
-    const classNames = getClassNames(getStyles!, { disabled, rowHeight });
+    const { disabled, rowHeight, isEditMode } = this.props;
+    const classNames = getClassNames(getStyles!, { disabled, rowHeight, isEditMode });
 
     if (this.props.isRenaming) {
       return (
         <div className={classNames.root}>
           <SectionTitleTextField
+            className={classNames.renameSectionTextField}
             id={this.props.id}
             placeHolder={this.props.title}
             rowHeight={this.props.rowHeight}
@@ -36,7 +37,7 @@ export class Section extends React.PureComponent<ISectionProps, ISectionState> {
       return (
         <div className={classNames.root}>
           <SectionTitleTextField
-            className={classNames.editTitleTextField}
+            className={classNames.addSectionTextField}
             id={this.props.id}
             placeHolder={this.props.title}
             rowHeight={this.props.rowHeight}
@@ -118,14 +119,7 @@ export class Section extends React.PureComponent<ISectionProps, ISectionState> {
   };
 
   private _onRenderOverflowItem(item: IOverflowSetItemProps): JSX.Element {
-    return (
-      <IconButton
-        menuIconProps={{ iconName: item.icon }}
-        onClick={item.onClick}
-        title={item.title}
-        ariaLabel={item.ariaLabel}
-      />
-    );
+    return <IconButton menuIconProps={{ iconName: item.icon }} onClick={item.onClick} title={item.title} ariaLabel={item.ariaLabel} />;
   }
 
   private _onRenderOverflowButton(overflowItems: IOverflowSetItemProps[] | undefined): JSX.Element {