+++ packages/experiments/src/components/CollapsibleSection/CollapsibleSectionTitle.styles.ts
@@ -14,8 +14,11 @@ export const getStyles = (props: IThemedProps<ICollapsibleSectionTitleProps>): I
         alignItems: 'center',
         background: 'none',
         border: 'none',
+        color: 'inherit',
         height: 24,
         margin: 0,
+        paddingLeft: 4 + (props.indent || 0) * 18,
+
         selectors: {
           ':hover': {
             background: theme.palette.neutralLight