+++ packages/experiments/src/components/Pagination/examples/Pagination.WithCombobox.Example.tsx
@@ -1,0 +1,39 @@
+import * as React from 'react';
+import { Pagination } from '@uifabric/experiments/lib/Pagination';
+
+export interface IPaginationBasicExampleState {
+  selectedPageIndex: number;
+}
+export class PaginationWithComboboxExample extends React.Component<{}, IPaginationBasicExampleState> {
+  constructor(props: {}) {
+    super(props);
+
+    this.state = { selectedPageIndex: 0 };
+  }
+
+  public render(): JSX.Element {
+    return (
+      <Pagination
+        selectedPageIndex={this.state.selectedPageIndex}
+        pageCount={28}
+        onPageChange={this.onPageChange}
+        withComboBox={true}
+        previousPageAriaLabel={'previous page'}
+        nextPageAriaLabel={'next page'}
+        firstPageAriaLabel={'first page'}
+        lastPageAriaLabel={'last page'}
+        pageAriaLabel={'page'}
+        firstPageIconProps={{ iconName: 'DoubleChevronLeft' }}
+        previousPageIconProps={{ iconName: 'ChevronLeft' }}
+        nextPageIconProps={{ iconName: 'ChevronRight' }}
+        lastPageIconProps={{ iconName: 'DoubleChevronRight' }}
+      />
+    );
+  }
+
+  private onPageChange = (index: number): void => {
+    this.setState({
+      selectedPageIndex: index
+    });
+  };
+}